+++ x/marker/types/authz.go
@@ -11,9 +11,11 @@ var (
 )
 
 // NewMarkerTransferAuthorization creates a new MarkerTransferAuthorization object.
-func NewMarkerTransferAuthorization(transferLimit sdk.Coins) *MarkerTransferAuthorization {
+func NewMarkerTransferAuthorization(transferLimit sdk.Coins, allowed []sdk.AccAddress) *MarkerTransferAuthorization {
+	allowedAddrs := toBech32Addresses(allowed)
 	return &MarkerTransferAuthorization{
 		TransferLimit: transferLimit,
+		AllowList:     allowedAddrs,
 	}
 }
 
@@ -26,6 +28,7 @@ func (a MarkerTransferAuthorization) MsgTypeURL() string {
 func (a MarkerTransferAuthorization) Accept(ctx sdk.Context, msg sdk.Msg) (authz.AcceptResponse, error) {
 	switch msg := msg.(type) {
 	case *MsgTransferRequest:
+		toAddress := msg.ToAddress
 		limitLeft, isNegative := a.DecreaseTransferLimit(msg.Amount)
 		if isNegative {
 			return authz.AcceptResponse{}, sdkerrors.ErrInsufficientFunds.Wrapf("requested amount is more than spend limit")
@@ -34,6 +37,21 @@ func (a MarkerTransferAuthorization) Accept(ctx sdk.Context, msg sdk.Msg) (authz
 		if limitLeft.IsZero() {
 			shouldDelete = true
 		}
+
+		isAddrExists := false
+		allowedList := a.GetAllowList()
+
+		for _, addr := range allowedList {
+			if addr == toAddress {
+				isAddrExists = true
+				break
+			}
+		}
+
+		if len(allowedList) > 0 && !isAddrExists {
+			return authz.AcceptResponse{}, sdkerrors.ErrUnauthorized.Wrapf("cannot send to %s address", toAddress)
+		}
+
 		return authz.AcceptResponse{Accept: true, Delete: shouldDelete, Updated: &MarkerTransferAuthorization{TransferLimit: limitLeft}}, nil
 	default:
 		return authz.AcceptResponse{}, sdkerrors.ErrInvalidType.Wrap("type mismatch")
@@ -48,10 +66,30 @@ func (a MarkerTransferAuthorization) ValidateBasic() error {
 	if !a.TransferLimit.IsAllPositive() {
 		return sdkerrors.ErrInvalidCoins.Wrap("spend limit cannot be negitive")
 	}
+	found := make(map[string]bool, 0)
+	for i := 0; i < len(a.AllowList); i++ {
+		if found[a.AllowList[i]] {
+			return ErrDuplicateEntry.Wrap("all allow list addresses must be unique")
+		}
+		found[a.AllowList[i]] = true
+	}
+
 	return nil
 }
 
 // DecreaseTransferLimit will return the decreased transfer limit and if it is negative
 func (a MarkerTransferAuthorization) DecreaseTransferLimit(amount sdk.Coin) (sdk.Coins, bool) {
 	return a.TransferLimit.SafeSub(amount)
+}
+
+func toBech32Addresses(allowed []sdk.AccAddress) []string {
+	if len(allowed) == 0 {
+		return nil
+	}
+
+	allowedAddrs := make([]string, len(allowed))
+	for i, addr := range allowed {
+		allowedAddrs[i] = addr.String()
+	}
+	return allowedAddrs
 }