+++ CHANGELOG.md
@@ -37,7 +37,13 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
-* nothing
+### Features
+
+* Added support to set a list of specific recipients allowed for send authorizations in the marker module [#1237](https://github.com/provenance-io/provenance/issues/1237).
+
+### Improvements
+
+* Added assess msg fees spec documentation [#1172](https://github.com/provenance-io/provenance/issues/1172).
 
 ---